+++ src/components/content/cars-page.tsx
@@ -1,6 +1,5 @@
 import SORTED_CARS, { FREE_CARS_COUNT } from "@/ir-data/utils/cars";
 import { ECarCategories } from "@/ir-data/utils/types";
-import { IR_URL } from "@/ir-data/utils/urls";
 import { useIr } from "@/store/ir";
 import { faCar } from "@fortawesome/free-solid-svg-icons";
 import ContentPage from "./content-page";
@@ -16,7 +15,6 @@ function CarsPage() {
         "Mark the cars you own and select your favorites. Use the wishlist to preview the content before you buy it."
       }
       freeCount={FREE_CARS_COUNT}
-      infoUrl={(id) => `${IR_URL.members}/web/shop/cars?carId=${id}`}
       myContent={myCars}
       skuIcon={faCar}
       tabs={ECarCategories}