+++ src/components/content/tracks-page.tsx
@@ -1,6 +1,5 @@
 import SORTED_TRACKS, { FREE_TRACKS_COUNT } from "@/ir-data/utils/tracks";
 import { ETrackCategories } from "@/ir-data/utils/types";
-import { IR_URL } from "@/ir-data/utils/urls";
 import { useIr } from "@/store/ir";
 import { faRoad } from "@fortawesome/free-solid-svg-icons";
 import ContentPage from "./content-page";
@@ -16,7 +15,6 @@ function TracksPage() {
         "Mark the tracks you own and select your favorites. Use the wishlist to preview the content before you buy it."
       }
       freeCount={FREE_TRACKS_COUNT}
-      infoUrl={(id) => `${IR_URL.members}/web/shop/tracks?trackId=${id}`}
       myContent={myTracks}
       skuIcon={faRoad}
       tabs={ETrackCategories}