+++ src/ir-data/utils/urls.ts
@@ -1,5 +1,4 @@
 export const IR_URL = {
-  members: "https://members-ng.iracing.com",
   image: "https://images-static.iracing.com",
   store: "https://members.iracing.com/membersite/member/store_r.jsp",
 };